feat: Adds result utils.
